+++ src/client/adapters/libsql/mod.ts
@@ -1,5 +1,5 @@
 export * from "./libsql-search-index.ts";
-export * from "./create-libsql-client.ts";
+export * from "./create-libsql-adapter.ts";
 export * from "./initialize-libsql-schema.ts";
 export * from "./libsql-client-base-options.ts";
 export * from "./hydrate-store-from-libsql.ts";